adding some additional leak sensors
can you tell i have had a leak in the past? ya, time to go all in
on these. right now, when a leak is detected there is automation
to send a critical notification to my phone /and/ alert over my
home sonos speakers with tts. in the future, i would also like to
integrate this with a relay on my water main to turn off water
pressure.