addiional sensor and some tweaks to existing ones
in this update, i'm adding the following new devices to my setup:

 zigbee2mqtt:
 * 1x contact sensor: aqara (xiaomi) MCCGQ11LM

 tasmota (+mqtt):
 * 1x switch: sonoff mini

... and removing:

 tasmota (+mqtt):
 * 1x switch: treatlife SS02

additionally, the ikea motion sensor that was previously being used in my office
has been moved to monitor for motion at the outside of the shed. with this data
and the new switch installed on the shed flood light, i am able to automatically
light up the pathway area in front of the shed (which includes the bbq pad when
cooking past dusk).

i swapped out the the treatlife SS02 previously being used for the flood light b/c
when it was off, it still provided a trickle of power to the led panel which dimmly
lit about 40% of the leds. while not an "issue" it did drive me crazy.