adding custom sonoff sv garage door opener
basic setup using a cover to open/close garage door through a
sonoff sv. currently the reed switch is not wired to the device,
so i am still using a wyzse sense contact sensors mounted on the
door to let home assistant know the state of the door. i will change
this once i come up with a good mounting mechanism for the wired
reed switch (and the sonoff sv + tasmota is already configured for
it).