rearrange ui to support garage door opener with confirmation
... i wanted to make sure there was a confirmation for the the
tap_action of opening the garage door since it's not something
that should be accidently clicked when everyone is away from the
house. i will also probably follow up with a notification on state
change that includes a photo from the wyze camera in the garage --
however, i still need to flash it with the rtsp supported firmware
to make that happen.