adding gosund wp2 plug for baby monitor (kitchen :: bedroom)
with this 2 outlet smart plug, i am using plug 1 (left) to switch
the baby monitor speaker. plug 2 (right) is set to be always on
for anything to be plugged into it ... standard kitchen outlet use.

to do this, with tasmota, i have the following rule to turn on the
relay when the device boots:

  `rule1 on system#boot do power2 on endon`