bug fix: garage heater would start when temp drop and door open
i was a bit too clever when setting up the garage heater automation
and missed the fact that if the door was open and the desired
heat temperature was reached, the heater would turn on. refactored
that logic to be much more simple. net effect, heater will be turned
off each time the door is opened and turned back on again when
closed. this vehicle door is only opened a couple times a day, so
this shouldn't be an issue