migrating to 433.92Mhz weather station sensors for some locations
i was having constant issues with zigbee devices dropping due to
power issues in the cold when trying to use them in the fridges
and freezers. sine i had a share software defined radio (sdr)
kicking around i figured it was worth a try with my existing
outdoor weather station. this data is pulled with rtl_433 on a
raspberry pi 4 and pushed to my mqtt server to be read within
home assistant. it's been much more stable and frankly, the
readings have been more accurate.