garage automation updates: lights and hvac temp normalization
there are two garage related automation changes in this commit:

1. new: turn garage ceiling lights on when garage vehicle door
opens after sunset. there is a condition to ensure the house door
hasn't recently been opened, which may indicate some is leaving
rather than returning.

2. update: evaluated the target vs. current temperature to set an
appropriate amount of time for the room temperature to normalize
before turning the hvac function back to heating when the door is
closed. previously, i was just waiting 6 minutes, but that wasn't
necessary unless it's really cold and introduced challenges if the
door was re-opened within that 6 minute delay.