updating scenes to remove invalid attributes
i hadnt noticed until this evening the following error message in
my logs:

```
The use of other attributes than device state attributes is
deprecated and will be removed in a future release. Invalid
attributes are ['brightness_pct', 'kelvin', 'transition'].
```

this commit cleans up those attributes. i may move these to scripts
to be able to utilize transitions, but will come back to that.